requirements.txt
openai-whisper
pydub
torch
numpy
